requirements.txt
asgiref==3.7.2
Django==4.2.6
django-stubs==4.2.6
django-stubs-ext==4.2.5
gunicorn==20.0.4
mypy==1.5.1
mypy-extensions==1.0.0
psycopg2-binary==2.9.5
sqlparse==0.4.4
tomli==2.0.1
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.8.0
whitenoise==6.5.0
djangorestframework==3.14.0
markdown==3.5.1
Pillow==10.1.0
drf-yasg==1.21.5
